+++ taxi_tilburg_backend/Models/Excel/LocationConnection.cs
@@ -4,6 +4,6 @@ public class LocationConnection
 {
     public int StartingPointId { get; set; }
     public int? TravelTimeInSeconds { get; set; }
-    public int? DistanceInKilometers { get; set; }
+    public decimal? DistanceInKilometers { get; set; }
     public int EndPointId { get; set; }
 }
