+++ taxi_tilburg_backend/Models/Responses/LocationConnectionItem.cs
@@ -5,5 +5,5 @@ public class LocationConnectionItem
     public int Id { get; set; }
     public string? Name { get; set; }
     public int TravelTimeInSeconds { get; set; }
-    public int DistanceInKilometers { get; set; }
+    public decimal DistanceInKilometers { get; set; }
 }
